Add a build using Ubuntu 14.04 (Trusty).
This is to validate that our build scripts work with our documented
minimum version of Perl (5.14).  This is the newest version of Ubuntu
on which Travis supports use of perl this old.

Contra the documentation, Trusty-based build workers’ default perl is
5.18, not 5.14.  Travis doesn’t support use of the `perl: VERSION`
build matrix parameter in a `language:c` build, and it doesn’t support
use of the `compiler:` parameter in a `language:perl` build.
The least terrible kludge I can come up with is to continue using
`language:c`, manually download and install the precompiled perl 5.14
binaries that Travis _would_ use for `perl:5.14`, and then manually
activate the appropriate perlbrew environment.  (We can’t use
$PERLBREW_ROOT/etc/bashrc because it’s, well, a bash script,
and our build scripts are _sh_ scripts.)

Since all this means we have to have an install script again anyway,
also use it to install perltidy and perlcritic when DISTCHECK=1,
add configuration files for both tools, and run perlcritic from the
distcheck-hook.  (This doesn’t *do* anything yet, of course, since
there aren’t any Perl scripts yet.)